Split out registration, added medication puller, and user medication service, authorized user table. Needs work to enable useInfo saving
