+++ src/main/java/com/vlol/config/WebSecurityConfig.java
@@ -88,7 +88,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .requestMatchers(EndpointRequest.toAnyEndpoint()).hasAuthority("admin")
 
                 .antMatchers("/admin/**")
-                    .hasAnyAuthority("participant", "provider", "agent", "admin")
+                    .hasAnyAuthority("participant", "provider", "admin")
                     .anyRequest()
                     .authenticated()
                     .and()
