+++ src/main/java/com/vlol/controller/QRController.java
@@ -19,5 +19,5 @@
 package com.vlol.controller;
 
 public class QRController {
-
+    
 }
