--- src/main/java/com/vlol/repository/UserMedListRepository.java
@@ -1,28 +1,0 @@
-/**
- * UserMedList Repository Interface.
- *
- * Java Runtime Environment (JRE) version used: 11.0.7
- * Java Development Kit (JDK) version used: 11.0.7
- *
- * Styling guide: Google Java Style Guide
- *     (https://google.github.io/styleguide/javaguide.html) and
- *     Code Conventions for the Java Programming Language (Oracle: Deprecated)
- *     (https://www.oracle.com/technetwork/java/javase/documentation/codeconvtoc-136057.html)
- *
- * @category  vlol
- * @package repository
- * @author Rob Garcia <rgarcia92@student.umgc.edu>
- * @license https://opensource.org/licenses/MIT The MIT License
- * @link      https://github.com/garciart/SWEN670
- * @copyright 2020 EMS Plus
- */
-package com.vlol.repository;
-
-import com.vlol.model.UserMedList;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserMedListRepository extends JpaRepository<UserMedList, Long> {
-
-}
