+++ src/main/java/com/vlol/repository/UserMedicationRepository.java
@@ -1,0 +1,42 @@
+/**
+ * UserMedication Repository Interface.
+ *
+ * Java Runtime Environment (JRE) version used: 11.0.7
+ * Java Development Kit (JDK) version used: 11.0.7
+ *
+ * Styling guide: Google Java Style Guide
+ *     (https://google.github.io/styleguide/javaguide.html) and
+ *     Code Conventions for the Java Programming Language (Oracle: Deprecated)
+ *     (https://www.oracle.com/technetwork/java/javase/documentation/codeconvtoc-136057.html)
+ *
+ * @category  vlol
+ * @package repository
+ * @author Rob Garcia <rgarcia92@student.umgc.edu>
+ * @license https://opensource.org/licenses/MIT The MIT License
+ * @link      https://github.com/garciart/SWEN670
+ * @copyright 2020 EMS Plus
+ */
+package com.vlol.repository;
+
+import com.vlol.model.UserMedication;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserMedicationRepository extends JpaRepository<UserMedication, Long> {
+    
+    @Query(value = "SELECT m FROM User_Medication m WHERE lower(m.brandName) LIKE lower(concat('%', :keyword, '%'))"
+            + "OR lower(m.genericName) LIKE lower(concat('%', :keyword, '%'))"
+            + "OR lower(m.drugAction) LIKE lower(concat('%', :keyword, '%'))", nativeQuery = true)
+    public List<UserMedication> findMedicationByKeyword(@Param("keyword") String keyword);
+    
+    @Query(value = "SELECT m FROM User u, UserMedication m WHERE u.email = :email")
+    public List<UserMedication> findMedicationByEmail(@Param("email") String email);
+    
+    
+}
