+++ src/main/resources/application.properties
@@ -22,7 +22,7 @@ spring.jpa.show-sql=true
 # ==============================================================
 # = Hibernate ddl auto (create, create-drop, update)
 # ==============================================================
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=none
 
 # ==============================================================
 # = The SQL dialect makes Hibernate generate better SQL for the chosen database
@@ -63,3 +63,9 @@ server.tomcat.accesslog.file-date-format=.yyyy-MM-dd
 management.endpoint.shutdown.enabled=true
 management.endpoint.info.enabled=true
 management.endpoints.web.exposure.include=*
+
+
+#logging.level.org.springframework.jdbc=DEBUG
+#logging.level.org.springframework.security=DEBUG
+#logging.level.org.springframework.web=DEBUG
+#logging.level.org.springframework.http=DEBUG
