+++ src/main/resources/templates/admin/authorized-user.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml"
+      layout:decorate="~{layouts/datatable-master}"
+      lang="en">
+    <head>
+        <title>Authorized User Manager</title>
+    </head>
+    <body>
+        <div layout:fragment="content" class="container">
+            <div class="row">
+                <div class="col-lg-12">
+                    <h1 class="pt-4">Authorized User Manager</h1>
+                    <form method="get" action="search-roles">
+                        <input type="text" name="keyword" pattern="[A-Za-z0-9\s\-._~:\/?#\[\]@!$&'()*+,;=]*" />&nbsp;&nbsp;
+                        <button type="submit" class="btn btn-info"><i class="fas fa-search"></i> Full Search</button>
+                    </form>
+                    <hr />
+                    <a href="/add-role" class="btn btn-success"><i class="fas fa-plus"></i> Add a User</a>
+                    <hr />
+                    <table id="list_table" class="table table-striped table-bordered dt-responsive display nowrap">
+                        <thead>
+                            <tr>
+                                <th data-priority="1">Email</th>
+                                <th data-priority="1">Name</th>
+                                <th data-priority="99" class="disabled-sorting">Actions</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <tr th:each="user : ${users}">
+                                <td th:text="${user.email}" class="align-middle">Email</td>
+                                <td th:text="${user.name}" class="align-middle">Name</td>
+                                <td class="text-nowrap">
+                                    <a th:href="@{/delete-authorized-user/{id}/{user.email}}" class="btn btn-danger"  onclick="return confirm('Are you sure you want to delete this user?')">
+                                        <i class="far fa-trash-alt"></i> Delete
+                                    </a>
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
+                    <br />
+                    <a href="/menu" class="btn btn-primary"><i class="fas fa-arrow-left"></i> Return to Menu</a>
+                    <br />
+                </div>
+            </div>
+        </div>
+    </body>
+</html>
